user-management/src/main/java/com/javaev/usermanagement/service/UserService.java
package com.javaev.usermanagement.service;

import com.javaev.usermanagement.model.User; 
import com.javaev.usermanagement.repository.UserRepository;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.UUID;

@Service
public class UserService {
    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;

    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    // Método para registrar usuarios
    public User registerUser(User user) {
        if (userRepository.findByEmail(user.getEmail()).isPresent()) {
            throw new RuntimeException("Email already exists");
        }

        // Encriptar la contraseña y guardar el usuario
        String encodedPassword = passwordEncoder.encode(user.getPassword());
        user.setPassword(encodedPassword);

        // Generar y asignar un token único
        user.setToken(UUID.randomUUID().toString());

        return userRepository.save(user);
    }
}
